BDOG-633 Split out non-update functions from DependencyDataUpdatingService
