+++ app/uk/gov/hmrc/servicedependencies/controller/admin/AdministrationController.scala
@@ -21,50 +21,61 @@ import play.api.Logger
 import play.api.libs.json.Json
 import play.api.mvc.ControllerComponents
 import uk.gov.hmrc.play.bootstrap.controller.BackendController
-import uk.gov.hmrc.servicedependencies.persistence.LocksRepository
+import uk.gov.hmrc.servicedependencies.persistence.{DependencyVersionRepository, LocksRepository, RepositoryLibraryDependenciesRepository}
 import uk.gov.hmrc.servicedependencies.service.DependencyDataUpdatingService
 
 import scala.concurrent.ExecutionContext
 
 @Singleton
 class AdministrationController @Inject()(
-    dependencyDataUpdatingService: DependencyDataUpdatingService
-  , locksRepository              : LocksRepository
-  , cc                           : ControllerComponents
+    dependencyDataUpdatingService          : DependencyDataUpdatingService
+  , locksRepository                        : LocksRepository
+  , repositoryLibraryDependenciesRepository: RepositoryLibraryDependenciesRepository
+  , dependencyVersionRepository            : DependencyVersionRepository
+  , cc                                     : ControllerComponents
   )(implicit ec: ExecutionContext
   ) extends BackendController(cc) {
 
-  def reloadLibraryDependenciesForAllRepositories(force: Option[Boolean] = None) = Action { implicit request =>
-    dependencyDataUpdatingService
-      .reloadCurrentDependenciesDataForAllRepositories(force = force.getOrElse(false))
-      .onFailure {
-        case ex => throw new RuntimeException("reload of dependencies failed", ex)
-      }
-    Accepted("reload started")
-  }
+  def reloadLibraryDependenciesForAllRepositories(force: Option[Boolean] = None) =
+    Action { implicit request =>
+      dependencyDataUpdatingService
+        .reloadCurrentDependenciesDataForAllRepositories(force = force.getOrElse(false))
+        .onFailure {
+          case ex => throw new RuntimeException("reload of dependencies failed", ex)
+        }
+      Accepted("reload started")
+    }
 
-  def reloadDependencyVersions = Action { implicit request =>
-    dependencyDataUpdatingService
-      .reloadLatestDependencyVersions
-      .onFailure {
-        case ex => throw new RuntimeException("reload of dependency versions failed", ex)
-      }
-    Accepted("reload started")
-  }
+  def reloadDependencyVersions =
+    Action { implicit request =>
+      dependencyDataUpdatingService
+        .reloadLatestDependencyVersions
+        .onFailure {
+          case ex => throw new RuntimeException("reload of dependency versions failed", ex)
+        }
+      Accepted("reload started")
+    }
 
-  def dropCollection(collection: String) = Action.async { implicit request =>
-    (collection match {
-       case "locks"    => locksRepository.clearAllData
-       case collection => dependencyDataUpdatingService.dropCollection(collection)
-     }
-    ).map(_ => Ok(s"$collection dropped"))
-  }
+  def dropCollection(collection: String) =
+    Action.async { implicit request =>
+      (collection match {
+         case "locks"                         => locksRepository.clearAllData
+         case "repositoryLibraryDependencies" => repositoryLibraryDependenciesRepository.clearAllData
+         case "dependencyVersions"            => dependencyVersionRepository.clearAllData
+         case other                           => sys.error(s"dropping $other collection is not supported")
+       }
+      ).map(_ => Ok(s"$collection dropped"))
+    }
 
-  def clearUpdateDates = Action.async { implicit request =>
-    dependencyDataUpdatingService.clearUpdateDates.map(rs => Ok(s"${rs.size} records updated"))
-  }
+  def clearUpdateDates =
+    Action.async { implicit request =>
+      repositoryLibraryDependenciesRepository
+        .clearUpdateDates
+        .map(rs => Ok(s"${rs.size} records updated"))
+    }
 
-  def mongoLocks() = Action.async { implicit request =>
-    locksRepository.getAllEntries.map(locks => Ok(Json.toJson(locks)))
-  }
+  def mongoLocks() =
+    Action.async { implicit request =>
+      locksRepository.getAllEntries.map(locks => Ok(Json.toJson(locks)))
+    }
 }
