+++ app/uk/gov/hmrc/servicedependencies/service/TeamDependencyService.scala
@@ -27,18 +27,18 @@ import scala.concurrent.{ExecutionContext, Future}
 
 @Singleton
 class TeamDependencyService @Inject()(
-  teamsAndReposConnector  : TeamsAndRepositoriesConnector
-, slugInfoRepository      : SlugInfoRepository
-, githubDepLookup         : DependencyDataUpdatingService
-, serviceConfigsConnector : ServiceConfigsConnector
-, slugDependenciesService : SlugDependenciesService
+  teamsAndReposConnector      : TeamsAndRepositoriesConnector
+, slugInfoRepository          : SlugInfoRepository
+, getMasterDependenciesService: GetMasterDependenciesService
+, serviceConfigsConnector     : ServiceConfigsConnector
+, slugDependenciesService     : SlugDependenciesService
 )(implicit ec: ExecutionContext
 ) {
 
   def findAllDepsForTeam(team: String)(implicit hc: HeaderCarrier): Future[Seq[Dependencies]] =
     for {
       (teamDetails, githubDeps) <- ( teamsAndReposConnector.getTeamDetails(team)
-                                   , githubDepLookup.getDependencyVersionsForAllRepositories
+                                   , getMasterDependenciesService.getDependencyVersionsForAllRepositories
                                    ).mapN { case (td, gh) => (td, gh) }
       libs                      =  teamDetails.libraries.flatMap(l => githubDeps.find(_.repositoryName == l))
       services                  =  teamDetails.services.flatMap(s => githubDeps.find(_.repositoryName == s))
@@ -55,7 +55,7 @@ class TeamDependencyService @Inject()(
 
   def dependenciesOfSlugForTeam(team: String, flag: SlugInfoFlag): Future[Map[String, Seq[Dependency]]] =
     for {
-      githubDeps <- githubDepLookup.getDependencyVersionsForAllRepositories
+      githubDeps <- getMasterDependenciesService.getDependencyVersionsForAllRepositories
       res        <- githubDeps.toList.traverse { githubDep =>
                       slugDependenciesService.curatedLibrariesOfSlug(githubDep.repositoryName, flag)
                         .map(_.map(githubDep.repositoryName -> _))
