+++ test/uk/gov/hmrc/servicedependencies/service/DependencyDataUpdatingServiceSpec.scala
@@ -58,18 +58,6 @@ class DependencyDataUpdatingServiceSpec
 
   private val timeForTest = Instant.now()
 
-  private val curatedDependencyConfig = CuratedDependencyConfig(
-    sbtPlugins = List( DependencyConfig(name = "plugin1", group = "uk.gov.hmrc" , latestVersion = None)
-                     , DependencyConfig(name = "plugin2", group = "uk.gov.hmrc" , latestVersion = None)
-                     )
-  , libraries  = List( DependencyConfig(name = "lib1"   , group= "uk.gov.hmrc"  , latestVersion = None)
-                     , DependencyConfig(name = "lib2"   , group= "uk.gov.hmrc"  , latestVersion = None)
-                     )
-  , others     = List( DependencyConfig(name = "sbt"    , group= "org.scala-sbt", latestVersion = None)
-                     )
-  )
-
-
   describe("reloadDependencyVersions") {
     it("should call the dependency version update function on the repository") {
       val boot = new Boot(CuratedDependencyConfig(
@@ -99,7 +87,11 @@ class DependencyDataUpdatingServiceSpec
     , repoLastUpdatedAt: Instant
     , shouldUpdate     : Boolean
     ) = {
-      val boot = new Boot(curatedDependencyConfig)
+      val boot = new Boot(CuratedDependencyConfig(
+        sbtPlugins = List.empty
+      , libraries  = List.empty
+      , others     = List.empty
+      ))
 
       val repositoryName = "repoXyz"
 
@@ -169,249 +161,6 @@ class DependencyDataUpdatingServiceSpec
       , repoLastUpdatedAt = Instant.EPOCH
       , shouldUpdate      = true
       )
-    }
-  }
-
-  describe("getDependencyVersionsForRepository") {
-    it("should return the current and latest library dependency versions for a repository") {
-      val boot = new Boot(curatedDependencyConfig)
-
-      val repositoryName = "repoXYZ"
-
-      when(boot.mockRepositoryLibraryDependenciesRepository.getForRepository(any()))
-        .thenReturn(Future.successful(
-          Some(MongoRepositoryDependencies(
-            repositoryName        = "repoXYZ"
-          , libraryDependencies   = Seq(
-                                      MongoRepositoryDependency(name = "lib1", group = "uk.gov.hmrc", currentVersion = Version("1.0.0"))
-                                    , MongoRepositoryDependency(name = "lib2", group = "uk.gov.hmrc", currentVersion = Version("2.0.0"))
-                                    )
-          , sbtPluginDependencies = Nil
-          , otherDependencies     = Nil
-          , updateDate            = timeForTest
-          ))
-        ))
-
-      val referenceLibraryVersions = Seq(
-        MongoDependencyVersion(name = "lib1", group = "uk.gov.hmrc", version = Version("1.1.0"))
-      , MongoDependencyVersion(name = "lib2", group = "uk.gov.hmrc", version = Version("2.1.0"))
-      )
-      when(boot.mockDependencyVersionRepository.getAllEntries)
-        .thenReturn(Future.successful(referenceLibraryVersions))
-
-      val optDependencies =
-        boot.dependencyUpdatingService.getDependencyVersionsForRepository(repositoryName).futureValue
-
-      verify(boot.mockRepositoryLibraryDependenciesRepository, times(1)).getForRepository(repositoryName)
-
-      optDependencies shouldBe Some(Dependencies(
-          repositoryName         = "repoXYZ"
-        , libraryDependencies    = Seq(
-                                     Dependency(name = "lib1", group = "uk.gov.hmrc", currentVersion = Version("1.0.0"), latestVersion = Some(Version("1.1.0")), bobbyRuleViolations = Nil)
-                                   , Dependency(name = "lib2", group = "uk.gov.hmrc", currentVersion = Version("2.0.0"), latestVersion = Some(Version("2.1.0")), bobbyRuleViolations = Nil)
-                                   )
-        , sbtPluginsDependencies = Nil
-        , otherDependencies      = Nil
-        , timeForTest
-        ))
-    }
-
-    it("should return the current and latest sbt plugin dependency versions for a repository") {
-      val boot = new Boot(curatedDependencyConfig)
-
-      val repositoryName = "repoXYZ"
-
-      when(boot.mockRepositoryLibraryDependenciesRepository.getForRepository(any()))
-        .thenReturn(Future.successful(
-          Some(MongoRepositoryDependencies(
-              repositoryName        = repositoryName
-            , libraryDependencies   = Nil
-            , sbtPluginDependencies = Seq(
-                                        MongoRepositoryDependency(name = "plugin1", group = "uk.gov.hmrc", currentVersion = Version(1, 0, 0))
-                                      , MongoRepositoryDependency(name = "plugin2", group = "uk.gov.hmrc", currentVersion = Version(2, 0, 0))
-                                      )
-            , otherDependencies     = Nil
-            , updateDate            = timeForTest
-            ))
-        ))
-
-      when(boot.mockDependencyVersionRepository.getAllEntries)
-        .thenReturn(Future.successful(Seq(
-            MongoDependencyVersion(name = "plugin1", group = "uk.gov.hmrc", version = Version("3.1.0"))
-          , MongoDependencyVersion(name = "plugin2", group = "uk.gov.hmrc", version = Version("4.1.0"))
-          )
-        ))
-
-      val optDependencies =
-        boot.dependencyUpdatingService.getDependencyVersionsForRepository(repositoryName).futureValue
-
-      verify(boot.mockRepositoryLibraryDependenciesRepository, times(1)).getForRepository(repositoryName)
-
-      optDependencies shouldBe Some(Dependencies(
-          repositoryName         = repositoryName
-        , libraryDependencies    = Nil
-        , sbtPluginsDependencies = Seq(
-                                     Dependency(name = "plugin1", group = "uk.gov.hmrc", currentVersion = Version("1.0.0"), latestVersion = Some(Version("3.1.0")), bobbyRuleViolations = Nil)
-                                   , Dependency(name = "plugin2", group = "uk.gov.hmrc", currentVersion = Version("2.0.0"), latestVersion = Some(Version("4.1.0")), bobbyRuleViolations = Nil)
-                                   )
-        , otherDependencies      = Nil
-        , lastUpdated            = timeForTest
-      ))
-    }
-
-    it("should return the current and latest external sbt plugin dependency versions for a repository") {
-      val boot = new Boot(CuratedDependencyConfig(
-          sbtPlugins = List( DependencyConfig(name = "internal-plugin", group = "uk.gov.hmrc", latestVersion = None)
-                           , DependencyConfig(name = "external-plugin", group = "uk.edu"     , latestVersion = None)
-                           )
-        , libraries  = Nil
-        , others     = Nil
-        ))
-
-      val repositoryName = "repoXYZ"
-
-      when(boot.mockRepositoryLibraryDependenciesRepository.getForRepository(any()))
-        .thenReturn(Future.successful(
-          Some(MongoRepositoryDependencies(
-              repositoryName        = repositoryName
-            , libraryDependencies   = Nil
-            , sbtPluginDependencies = Seq(
-                                        MongoRepositoryDependency(name = "internal-plugin", group = "uk.gov.hmrc", currentVersion = Version("1.0.0"))
-                                      , MongoRepositoryDependency(name = "external-plugin", group = "uk.edu"     , currentVersion = Version("2.0.0"))
-                                      )
-            , otherDependencies     = Nil
-            , updateDate            = timeForTest
-            ))
-        ))
-
-      when(boot.mockDependencyVersionRepository.getAllEntries)
-        .thenReturn(
-          Future.successful(Seq(
-              MongoDependencyVersion(name = "internal-plugin", group = "uk.gov.hmrc", version = Version("3.1.0"))
-            , MongoDependencyVersion(name = "external-plugin", group = "uk.edu"     , version = Version("11.22.33"))
-            ))
-        )
-
-      val optDependencies =
-        boot.dependencyUpdatingService.getDependencyVersionsForRepository(repositoryName).futureValue
-
-      optDependencies shouldBe Some(Dependencies(
-          repositoryName         = repositoryName
-        , libraryDependencies    = Nil
-        , sbtPluginsDependencies = Seq(
-                                     Dependency(name = "internal-plugin", group = "uk.gov.hmrc", currentVersion = Version("1.0.0"), latestVersion = Some(Version("3.1.0"))   , bobbyRuleViolations = Nil)
-                                   , Dependency(name = "external-plugin", group = "uk.edu"     , currentVersion = Version("2.0.0"), latestVersion = Some(Version("11.22.33")), bobbyRuleViolations = Nil)
-                                   )
-        , otherDependencies      = Nil
-        , lastUpdated            = timeForTest
-        ))
-
-      verify(boot.mockRepositoryLibraryDependenciesRepository, times(1)).getForRepository(repositoryName)
-    }
-
-    it("test for none") {
-      val boot = new Boot(curatedDependencyConfig)
-
-      val repositoryName = "repoXYZ"
-
-      when(boot.mockRepositoryLibraryDependenciesRepository.getForRepository(any()))
-        .thenReturn(Future.successful(None))
-
-      when(boot.mockDependencyVersionRepository.getAllEntries)
-        .thenReturn(Future.successful(Nil))
-
-      val maybeDependencies =
-        boot.dependencyUpdatingService.getDependencyVersionsForRepository(repositoryName).futureValue
-
-      verify(boot.mockRepositoryLibraryDependenciesRepository, times(1)).getForRepository(repositoryName)
-
-      maybeDependencies shouldBe None
-    }
-  }
-
-  describe("getDependencyVersionsForAllRepositories") {
-    it("should return the current and latest library, sbt plugin and other dependency versions for all repositories") {
-      val boot = new Boot(curatedDependencyConfig)
-
-      when(boot.mockRepositoryLibraryDependenciesRepository.getAllEntries)
-        .thenReturn(Future.successful(Seq(
-            MongoRepositoryDependencies(
-                repositoryName        = "repo1"
-              , libraryDependencies   = Seq(
-                                          MongoRepositoryDependency(name = "lib1", group = "uk.gov.hmrc", currentVersion = Version("1.1.0"))
-                                        , MongoRepositoryDependency(name = "lib2", group = "uk.gov.hmrc", currentVersion = Version("1.2.0"))
-                                        )
-              , sbtPluginDependencies = Seq(
-                                          MongoRepositoryDependency(name = "plugin1", group = "uk.gov.hmrc", currentVersion = Version("10.1.0"))
-                                        , MongoRepositoryDependency(name = "plugin2", group = "uk.gov.hmrc", currentVersion = Version("10.2.0"))
-                                        )
-              , otherDependencies     = Seq(
-                                          MongoRepositoryDependency(name = "sbt", group = "org.scala-sbt", currentVersion = Version("0.13.1"))
-                                        )
-              , updateDate            = timeForTest
-              )
-          , MongoRepositoryDependencies(
-              repositoryName        = "repo2"
-            , libraryDependencies   = Seq(
-                                        MongoRepositoryDependency(name = "lib1", group = "uk.gov.hmrc", currentVersion = Version("2.1.0"))
-                                      , MongoRepositoryDependency(name = "lib2", group = "uk.gov.hmrc", currentVersion = Version("2.2.0"))
-                                      )
-            , sbtPluginDependencies = Seq(
-                                        MongoRepositoryDependency(name = "plugin1", group = "uk.gov.hmrc", currentVersion = Version("20.1.0"))
-                                      , MongoRepositoryDependency(name = "plugin2", group = "uk.gov.hmrc", currentVersion = Version("20.2.0"))
-                                      )
-            , otherDependencies     = Seq(
-                                        MongoRepositoryDependency(name = "sbt", group = "org.scala-sbt", currentVersion = Version("0.13.2"))
-                                      )
-            , updateDate            = timeForTest
-            )
-        )))
-
-      when(boot.mockDependencyVersionRepository.getAllEntries)
-        .thenReturn(Future.successful(Seq(
-            MongoDependencyVersion(name = "lib1"   , group = "uk.gov.hmrc"  , version = Version("3.0.0"))
-          , MongoDependencyVersion(name = "lib2"   , group = "uk.gov.hmrc"  , version = Version("4.0.0"))
-          , MongoDependencyVersion(name = "plugin1", group = "uk.gov.hmrc"  , version = Version("30.0.0"))
-          , MongoDependencyVersion(name = "plugin2", group = "uk.gov.hmrc"  , version = Version("40.0.0"))
-          , MongoDependencyVersion(name = "sbt"    , group = "org.scala-sbt", version = Version("100.10.1"))
-          )))
-
-      val maybeDependencies = boot.dependencyUpdatingService.getDependencyVersionsForAllRepositories.futureValue
-
-      verify(boot.mockRepositoryLibraryDependenciesRepository, times(1)).getAllEntries
-
-      maybeDependencies should contain theSameElementsAs Seq(
-          Dependencies(
-            repositoryName = "repo1"
-          , libraryDependencies = Seq(
-              Dependency(name = "lib1", group = "uk.gov.hmrc", currentVersion = Version("1.1.0"), latestVersion = Some(Version("3.0.0")), bobbyRuleViolations = Nil)
-            , Dependency(name = "lib2", group = "uk.gov.hmrc", currentVersion = Version("1.2.0"), latestVersion = Some(Version("4.0.0")), bobbyRuleViolations = Nil)
-            )
-          , sbtPluginsDependencies = Seq(
-              Dependency(name = "plugin1", group = "uk.gov.hmrc", currentVersion = Version("10.1.0"), latestVersion = Some(Version("30.0.0")), bobbyRuleViolations = Nil)
-            , Dependency(name = "plugin2", group = "uk.gov.hmrc", currentVersion = Version("10.2.0"), latestVersion = Some(Version("40.0.0")), bobbyRuleViolations = Nil)
-            )
-          , otherDependencies = Seq(
-              Dependency(name = "sbt", group = "org.scala-sbt", currentVersion = Version("0.13.1"), latestVersion = Some(Version("100.10.1")), bobbyRuleViolations = Nil)
-            )
-          , timeForTest
-          )
-        , Dependencies(
-            repositoryName = "repo2"
-          , libraryDependencies = Seq(
-              Dependency(name = "lib1", group = "uk.gov.hmrc", currentVersion = Version("2.1.0"), latestVersion = Some(Version("3.0.0")), bobbyRuleViolations = Nil),
-              Dependency(name = "lib2", group = "uk.gov.hmrc", currentVersion = Version("2.2.0"), latestVersion = Some(Version("4.0.0")), bobbyRuleViolations = Nil)
-            )
-          , sbtPluginsDependencies = Seq(
-              Dependency(name = "plugin1", group = "uk.gov.hmrc", currentVersion = Version("20.1.0"), latestVersion = Some(Version("30.0.0")), bobbyRuleViolations = Nil),
-              Dependency(name = "plugin2", group = "uk.gov.hmrc", currentVersion = Version("20.2.0"), latestVersion = Some(Version("40.0.0")), bobbyRuleViolations = Nil)
-            )
-          , otherDependencies = Seq(
-              Dependency(name = "sbt", group = "org.scala-sbt", currentVersion = Version("0.13.2"), latestVersion = Some(Version("100.10.1")), bobbyRuleViolations = Nil)
-            )
-          , timeForTest
-          )
-        )
     }
   }
